+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **fix:** change eslint (no-param-reassign) rule
 - **add:** hapi (13.x.x) as peer dependency
 - **add:** joi (8.1.0) as dependency
+- **add:** joi based validation for plugin options
 
 ### [v0.1.2] - 2016-05-15
 - **doc:** fix typo